Refactor API routes, service and base URLs. Add announcements docs

+++ api/services/announcements.js
@@ -5,8 +5,8 @@ const config = require("../config");
 const announcementProperties = [
 	"title",
 	"content",
-	"valid_until",
-	"personnel_id",
+	"validUntil",
+	"personnelId",
 ];
 
 function validateAnnouncement(announcement) {
@@ -16,24 +16,24 @@ function validateAnnouncement(announcement) {
 	const datetimeRegex =
 		/^(((\d{4})-([01]\d)-(0[1-9]|[12]\d|3[01])) (([01]\d|2[0-3]):([0-5]\d):([0-5]\d)))$/m;
 
-	// check if valid_until is in correct format
-	if (!datetimeRegex.test(announcement.valid_until)) {
+	// check if validUntil is in correct format
+	if (!datetimeRegex.test(announcement.validUntil)) {
 		throw new Error(
 			JSON.stringify({
-				statusMessage: "Valid until property invalid format",
+				statusMessage: "validUntil property invalid format",
 				statusCode: 400,
 			})
 		);
 	}
 
-	// check if valid_until is greater than current date
+	// check if validUntil is greater than current date
 	const currentDate = new Date();
-	const validUntilDate = new Date(announcement.valid_until);
+	const validUntilDate = new Date(announcement.validUntil);
 	if (validUntilDate < currentDate) {
 		throw new Error(
 			JSON.stringify({
 				statusMessage:
-					"Valid until date should be greater than current date",
+					"validUntil date should be greater than current date",
 				statusCode: 400,
 			})
 		);
@@ -66,6 +66,15 @@ async function getById(id) {
 	const rows = await db.query("SELECT * FROM Announcements WHERE id=?", [id]);
 	const data = helper.emptyOrRows(rows);
 
+	if (data.length === 0) {
+		throw new Error(
+			JSON.stringify({
+				statusMessage: "Announcement not found",
+				statusCode: 404,
+			})
+		);
+	}
+
 	return {
 		data,
 		response: { message: `Successfully fetched data`, statusCode: 200 },
@@ -75,8 +84,28 @@ async function getById(id) {
 async function create(announcement) {
 	validateAnnouncement(announcement);
 
-	const result = await db.query("INSERT INTO Announcements SET=?", [
-		announcement,
+	const announcementExists = await db.query(
+		"SELECT IF(COUNT(*)=0,0,1) result FROM Announcements WHERE id=?",
+		[id]
+	);
+
+	if (announcementExists[0].result) {
+		throw new Error(
+			JSON.stringify({
+				statusMessage: "Announcement with this id already exists",
+				statusCode: 409,
+			})
+		);
+	}
+
+	const result = await db.query("INSERT INTO Announcements SET ?", [
+		{
+			Id: announcement.id ? announcement.id : null,
+			Title: announcement.title,
+			Content: announcement.content,
+			Valid_until: announcement.validUntil,
+			Personnel_id: announcement.personnelId,
+		},
 	]);
 
 	if (result.affectedRows) {
@@ -97,8 +126,28 @@ async function create(announcement) {
 async function update(id, announcement) {
 	validateAnnouncement(announcement);
 
+	const announcementExists = await db.query(
+		"SELECT IF(COUNT(*)=0,0,1) result FROM Announcements WHERE id=?",
+		[id]
+	);
+
+	if (!announcementExists[0].result) {
+		throw new Error(
+			JSON.stringify({
+				statusMessage: "Announcement with this id does not exist",
+				statusCode: 404,
+			})
+		);
+	}
+
 	const result = await db.query("UPDATE Announcements SET ? WHERE id=?", [
-		announcement,
+		{
+			Id: announcement.id ? announcement.id : id,
+			Title: announcement.title,
+			Content: announcement.content,
+			Valid_until: announcement.validUntil,
+			Personnel_id: announcement.personnelId,
+		},
 		id,
 	]);
 
@@ -118,6 +167,20 @@ async function update(id, announcement) {
 }
 
 async function remove(id) {
+	const announcementExists = await db.query(
+		"SELECT IF(COUNT(*)=0,0,1) result FROM Announcements WHERE id=?",
+		[id]
+	);
+
+	if (!announcementExists[0].result) {
+		throw new Error(
+			JSON.stringify({
+				statusMessage: "Announcement with this id does not exist",
+				statusCode: 404,
+			})
+		);
+	}
+
 	const result = await db.query("DELETE FROM Announcements WHERE id=?", [id]);
 
 	if (result.affectedRows) {
@@ -128,7 +191,7 @@ async function remove(id) {
 	} else {
 		throw new Error(
 			JSON.stringify({
-				message: "Failed to delete announcement",
+				message: "Announcement could not be deleted",
 				statusCode: 500,
 			})
 		);
